Decompose functions in accumulator
